+++ site/html/src/delete_user.php
@@ -4,9 +4,13 @@ require_once("../config/session.php");
 
 session_start();
 
+print_r($_SESSION);
+
 //check if session has expired
 $mySession = new MySession();
 $mySession->isLoginSessionExpired();
+
+
 
 
 if (!isset($_SESSION['email'])) {
@@ -14,7 +18,7 @@ if (!isset($_SESSION['email'])) {
 } else {
     $db = new MyDB();
 
-    if($_SESSION['isAdmin'] =!true)
+    if($_SESSION['isAdmin'] !== true)
     {
         header('Location: ./mailbox.php');
 
@@ -24,11 +28,19 @@ if (!isset($_SESSION['email'])) {
 $type=0;
 
 $user=filter_var($_GET['user'], FILTER_SANITIZE_STRING);
+$token = filter_var($_GET['token'], FILTER_SANITIZE_STRING);
+
 
 if(isset($user))
 {
+
     $db = new MyDB();
-    $db->deleteUser($user);
+    if( $_SESSION ['token'] === $token) {
+
+        $db->deleteUser($user);
+    }
+
+
     header('Location: ./users.php');
